Version 2.0.2
Accounted for a problem that caused new layers from biological records
tool to fail to load if a previous layer generated by the tool is
manually removed from the map. This can be fixed clearing all layers
using tool's button, so a new error trap and associated message added to
indicate this.