Version 2.0.3
Added facility for the Biological Records tool to generate two extra
attributes for atlas layers - 'richness' and 'taxa'. Richness is a count
of the number of taxa which contribute to the records and 'taxa' is a
string of all taxa names separated by the hash symbol. The latter allows
you to use QGIS categorical styling to create coincidence maps for
biological records.