Version 2.0.4
Added ability to store environment options in a file that is external to
the plugin folder. The 'internal' environment options folder gets
deleted/replaced whenever the plugin is updated, so this is to overcome
that problem. Added support for NBN username and password environment
options. Fixed a bug with generation of values for 'taxa' column of
atlas maps.